Implement UI feedback and require explicit confirmation of new room
This matches the previous workflow, but with better feedback. It also
gives an easy place to allow for more custom configuration in the
future. Unfortunately, XEP-0045 is completely idiotic, and that requires
data forms. It will be a long long time before we have the UI to present
those, if ever.