Delay topmsg calculation until access
This is in preparation for the next change which requires marking the
value for recompute to be cheap. It is also so we can eventually move it
to some appropriate signal instead of being something everyone must
remember to call.

We recompute on access and when switching direction. Unfortunately, it
is still somewhat annoyingly stateful in that regard. Code is also still
sprawled all over the place it should someday get consolidated into the
mainwin.

FIXME: This also doesn't give exactly the same result as the old one in
all cases. Thousands of nexts in a row. I think this is okay, but we'll
see. The intent is to avoid recalculating topmsg O(N) times when
expunging N messages.