Replace gval_default with a string
This is in preparation for making C <-> perl interface marshall
everything as strings because that's how they're used anyway. (Perl
variables are only accessed from C to be get and set from the user
interface which reads/displays as strings. Internally, the variable is
free to use whatever it wants.)