Kill owl_context_is_interactive checks
This eliminates the difference between OWL_CTX_INTERACTIVE and
OWL_CTX_STARTUP. The two functions which were sensitive used it to
suppress some output. If called from a command, this only makes a
difference if run from .owl/init.pl. This is a behavior change, but a
boring one, and most other commands will print output.

The only one that is called outside a command is
owl_function_loadloginsubs. That's called at
owl_zephyr_finish_initialization which runs after the event loop has
started and we're in interactive context. So the check doesn't do
anything in that case.