Actually enforce contexts on all commands
Now that OWL_CTX_ALL and OWL_CTX_INTERACTIVE are equivalent, this
results in no behavior change: all commands that were unchecked are
still always runnable. With the exception of yes/no which were marked as
OWL_CTX_RECV. They run a command callback, so we want to be sure of
their contexts anyway. We can flip them to OWL_CTX_INTERACTIVE if
needbe. (And likewise we probably want to flip some of these to
OWL_CTX_RECV at some point...)