Signal context deactivating after the new one has taken effect
I think this is what you want? None of the current uses care, but it
seems a blocking context's callback should run in the parent context.
Possible it should be a different callback instead.

If a context is a UI-level function call, the deactivate callback should
be a continuation, and then you want it to run in the previous context,
so that makes some sense. Arguably it should also take an argument,
which would make uses of OWL_CTX_BLOCKING less awkward.