Tweak Jabber message attributes and display
This might be a little controversial and needs feedback. I think it's
cleaner. A lot is motivated by what fields are and aren't meaningful in
GTalk, as it's probably the most used Jabber deployment, especially for
MUCs.

- from/to message fields now always refer to the Jabber-level from/to
  attributes.
- sender/recipient are the logical sender and recipient. nick for
  anonymous MUC, JID for nonymous MUC. (Flipped from email, but oh
  well.)
- don't display resources in personal_context. They are almost never
  meaningful, and they were already not in sender/recipient fields.
- in nonymous MUCs, drop nick in favor of sender. This reflects GTalk
  picking a meaningless nick in MUCs. Do this for personals to MUC
  members too. It does, however, not match Pidgin.
- the long_sender is always the name of the logical sender if known.
  Otherwise we'll fall back to the full JID and whatnot.

This does have the slight oddity that you never see the nick of someone
in a nonymous MUC if you also have a friendly name for them. (Currently
only checking the roster, but we should implement XEP-0172.) I think
this is fine. It's all still in :info if you want it.